parsing_agent/requirements.txt
openai>=1.60.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.32.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pandas>=2.2.0
pyyaml>=6.0.0
tiktoken>=0.8.0
python-dotenv>=1.0.0
schedule>=1.2.0
lxml>=5.0.0
html2text>=2024.2.0
